Moving TypeBasedSagaMetaModel capabilities into SagaMetadata.
